fix feat, struct
