+++ atlas-ui-kit/src/styles/variables.css
@@ -29,8 +29,6 @@
 	--stroke: #525252;
 	--stroke-inv: #a7a7a7;
 
-	--disabled: #525252
-
-	--border-radius-block: 32px;
+	--disabled: #525252 --border-radius-block: 32px;
 	--border-radius: 56px;
 }
