+++ client/src/app/main.tsx
@@ -7,8 +7,8 @@ import { Provider } from 'react-redux';
 
 createRoot(document.getElementById('root')!).render(
 	<StrictMode>
-			<Provider store={store}>
-				<App />
-			</Provider>
+		<Provider store={store}>
+			<App />
+		</Provider>
 	</StrictMode>
 );
