+++ client/src/features/model3d/types.ts
@@ -1,0 +1,14 @@
+import type { FileInfo } from "@shared/types/fileTypes";
+
+export type Model3DState = {
+  list: FileInfo[];
+  current?: FileInfo | null;
+  modelUrl?: string | null;
+  loadingList: boolean;
+  loadingModel: boolean;
+  uploading: boolean;
+  deleting: boolean;
+  error?: string | null;
+};
+
+export type ModelProps = { modelUrl?: string | null; height?: number };
