+++ client/src/shared/config/apiBase.ts
@@ -14,3 +14,5 @@ export const WS_BASE_URL =
       return `${wsProto}${window.location.host}`;
     }
   })();
+
+// render feat
