+++ client/src/shared/types/fileTypes.ts
@@ -1,0 +1,6 @@
+export type FileInfo = {
+  objectKey: string;
+  size?: number;
+  contentType?: string;
+  createdAt?: string;
+};
